BHBC-2224: Rename attachments endpoint to artifacts
