+++ api/src/paths/dwc/submission/{datasetId}/artifacts.ts
